Refactor DataTableApiResponse to use PagerTableApiData for improved code organization
